requirements.txt
streamlit
python-docx
pywin32